+++ tempCodeRunnerFile.py
@@ -1,0 +1,3 @@
+    # slash = await bot.tree.sync()
+    # print(f"目前登入身份 --> {bot.user}")
+    # print(f"載入 {len(slash)} 個斜線指令")
